+++ src/pydidas/gui/frames/data_browsing_frame.py
@@ -37,7 +37,7 @@ from silx.gui.hdf5 import H5Node
 from silx.gui.hdf5.Hdf5Item import Hdf5Item
 from silx.gui.hdf5.Hdf5Node import Hdf5Node
 
-from pydidas.core import Dataset, Parameter, ParameterCollection
+from pydidas.core import Dataset, Parameter, ParameterCollection, get_generic_parameter
 from pydidas.core.constants import (
     ALIGN_TOP_RIGHT,
 )
@@ -56,6 +56,7 @@ from pydidas.gui.frames.builders.data_browsing_frame_builder import (
 )
 from pydidas.widgets.framework import BaseFrame, PydidasWindow
 from pydidas.widgets.misc import ReadOnlyTextWidget
+from pydidas.widgets.parameter_config.param_io_widget_file import ParamIoWidgetFile
 from pydidas.widgets.windows import Hdf5BrowserWindow
 
 
@@ -79,7 +80,7 @@ class DataBrowsingFrame(BaseFrame, AssociatedFileMixin):
             name="[ASCII] Use data column as x values:",
             allow_None=True,
             choices=[None],
-        )
+        ),
     )
     params_not_to_restore = ["xcol"]
 
@@ -94,15 +95,17 @@ class DataBrowsingFrame(BaseFrame, AssociatedFileMixin):
         self.set_default_params()
 
     def connect_signals(self) -> None:
-        """
-        Connect all required signals and slots between widgets and class methods.
-        """
+        """Connect all required signals and slots between widgets and class methods."""
+        # Signals to handle new file selections:
         self._widgets["explorer"].sig_new_file_selected.connect(self.__file_selected)
-        self._widgets["filename"].returnPressed.connect(self.__filename_input)
-        self._widgets["filename"].editingFinished.connect(
-            self.__set_filename_input_text
-        )
+        _file_io = self._widgets["filename"]._io_lineedit
+        _file_io.returnPressed.connect(self.__filename_input)
+        _file_io.editingFinished.connect(self.__restore_filename_display_text)
+        self._widgets["filename"].sig_drop_accepted.connect(self.__filename_input)
+
+        # Signals for special file-type handling (e.g. ASCII, HDF5)
         self.param_widgets["xcol"].sig_new_value.connect(self.__display_ascii_data)
+        self._widgets["button_ascii_metadata"].clicked.connect(self.__display_metadata)
         self._widgets["hdf5_dataset_selector"].sig_new_dataset_selected.connect(
             self.__display_hdf5_dataset
         )
@@ -112,7 +115,6 @@ class DataBrowsingFrame(BaseFrame, AssociatedFileMixin):
         self._widgets["hdf5_dataset_selector"].sig_request_hdf5_browser.connect(
             self.__inspect_hdf5_tree
         )
-        self._widgets["button_ascii_metadata"].clicked.connect(self.__display_metadata)
 
     def build_frame(self) -> None:
         """
@@ -121,16 +123,30 @@ class DataBrowsingFrame(BaseFrame, AssociatedFileMixin):
         for _method, _args, _kwargs in DATA_BROWSING_FRAME_BUILD_CONFIG:
             _widget_creation_method = getattr(self, _method)
             _widget_creation_method(*_args, **_kwargs)
+        # explicitly add the widget here to because the generic widget
+        # creation do not accept direct args
+        self.add_any_widget(
+            "filename",
+            ParamIoWidgetFile(get_generic_parameter("filename")),
+            gridPos=(0, 1, 1, 2),
+            parent_widget="plot_header",
+        )
         self.add_any_widget(
             "splitter",
             create_splitter(
-                self._widgets["browser"],
+                self._widgets["browser_canvas"],
                 self._widgets["viewer_and_filename"],
                 self.width(),
             ),
         )
         _viewer_layout = self._widgets["viewer_and_filename"].layout()
         _viewer_layout.setRowStretch(_viewer_layout.rowCount() - 1, 1)
+
+    def finalize_ui(self) -> None:
+        """Finalize the UI initialization."""
+        super().finalize_ui()
+        self.add_param(self._widgets["filename"]._linked_param)
+        self._widgets["filename"]._button.setVisible(False)
 
     @QtCore.Slot(int)
     def frame_activated(self, index: int) -> None:
@@ -152,6 +168,8 @@ class DataBrowsingFrame(BaseFrame, AssociatedFileMixin):
             for _window in (self.__browser_window, self.__metadata_window):
                 if _window is not None and _window.isVisible():
                     _window.hide()
+        if index == self.frame_index:
+            self._widgets["explorer"].check_root_up_to_date()
 
     @QtCore.Slot(str)
     def __file_selected(self, filename: str) -> None:
@@ -170,7 +188,7 @@ class DataBrowsingFrame(BaseFrame, AssociatedFileMixin):
             get_extension(filename) not in self.__supported_extensions
             or not Path(filename).is_file()
         ):
-            self.__set_filename_input_text()
+            self.__restore_filename_display_text()
             self._widgets["viewer"].set_data(None)
             if self.__browser_window is not None:
                 self.__browser_window.hide()
@@ -182,7 +200,7 @@ class DataBrowsingFrame(BaseFrame, AssociatedFileMixin):
         if self.__metadata_window is not None:
             self.__metadata_window.hide()
         self._widgets["viewer"].set_data(None)
-        self.__set_filename_input_text()
+        self.__restore_filename_display_text()
         self._widgets["ascii_widgets"].setVisible(self.ascii_file)
         self._widgets["hdf5_dataset_selector"].setVisible(self.hdf5_file)
         self._widgets["configure_binary_decoding"].set_new_filename(filename)
@@ -194,17 +212,22 @@ class DataBrowsingFrame(BaseFrame, AssociatedFileMixin):
             _data = import_data(filename)
             self.__display_dataset(_data)
 
+    @QtCore.Slot()
     def __filename_input(self):
-        self.__file_selected(self._widgets["filename"].text())
+        """Select the current file based on a user filename input."""
+        self.__file_selected(self._widgets["filename"].current_text)
 
-    def __set_filename_input_text(self):
+    @QtCore.Slot()
+    def __restore_filename_display_text(self):
         """
-        Update filename input text only if it does not already match the current
-        filename. This is to prevent the text cursor from jumping to the end of
-        the filename when using the filename input field.
+        Restore the filename displayed in the UI.
+
+        This method updates the filename displayed in the UI in case of
+        unconfirmed user edits to assure that the displayed data and
+        filename are synchronized.
         """
-        if self._widgets["filename"].text() != self.current_filename:
-            self._widgets["filename"].setText(self.current_filename)
+        if self._widgets["filename"].current_text != self.current_filename:
+            self._widgets["filename"].update_display_value(self.current_filename)
 
     def __open_hdf5_file(self) -> None:
         """Process the input file and check whether it is a hdf5 file."""
@@ -227,7 +250,7 @@ class DataBrowsingFrame(BaseFrame, AssociatedFileMixin):
                 pass
             self.__open_file = None
             self._widgets["hdf5_dataset_selector"].clear()
-            self._widgets["filename"].setText("")
+            self._widgets["filename"].update_display_value("")
             self.current_filename = ""
             raise FileReadError(catcher.exception_message)