+++ src/pydidas/widgets/scroll_area.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2024 - 2025, Helmholtz-Zentrum Hereon
+# Copyright 2024 - 2026, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -16,23 +16,26 @@
 # along with Pydidas. If not, see <http://www.gnu.org/licenses/>.
 
 """
-Module with the ScrollAreaclass, a QScrollArea implementation with convenience
-calling arguments for simplified formatting.
+Module with the ScrollArea class, a QScrollArea implementation with
+convenience calling arguments for simplified formatting.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2024 - 2025, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2024 - 2026, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
 __all__ = ["ScrollArea"]
 
 
+from typing import Any
+
 from qtpy.QtCore import QSize, Slot
-from qtpy.QtWidgets import QApplication, QFrame, QScrollArea
+from qtpy.QtWidgets import QFrame, QScrollArea, QWidget
 
 from pydidas.core.constants import POLICY_EXP_EXP
 from pydidas.core.utils import apply_qt_properties
+from pydidas_qtcore import PydidasQApplication
 
 
 class ScrollArea(QScrollArea):
@@ -43,27 +46,28 @@ class ScrollArea(QScrollArea):
     ----------
     parent : QWidget, optional
         The parent widget. The default is None.
-    **kwargs : dict
-        Any additional keyword arguments. All Qt attributes with a setAttribute
-        method are valid keywords.
+    **kwargs : Any
+        Any additional keyword arguments. All Qt attributes with a
+        setAttribute method are valid keywords.
     """
 
-    init_kwargs = ["resize_to_widget_width"]
+    init_kwargs = ["resize_to_widget_width", "embedded_widget"]
 
-    def __init__(self, parent=None, **kwargs):
-        QScrollArea.__init__(self, parent)
+    def __init__(self, parent: QWidget | None = None, **kwargs: Any) -> None:
+        QScrollArea.__init__(self, parent)  # type: ignore[arg-type]
         kwargs["widgetResizable"] = True
         kwargs["autoFillBackground"] = True
         kwargs["sizePolicy"] = POLICY_EXP_EXP
         kwargs["frameShape"] = QFrame.NoFrame
-        apply_qt_properties(self, **kwargs)
-        self.__scrollbar_width = QApplication.instance().scrollbar_width
+        apply_qt_properties(self, **kwargs)  # type: ignore[arg-type]
+        _app = PydidasQApplication.instance()
+        self.__scrollbar_width = _app.scrollbar_width  # type: ignore[attr-defined]
         if kwargs.get("resize_to_widget_width", False):
-            QApplication.instance().sig_font_metrics_changed.connect(
+            _app.sig_font_metrics_changed.connect(  # type: ignore[attr-defined]
                 self.force_width_from_size_hint
             )
 
-    def sizeHint(self):
+    def sizeHint(self) -> QSize:
         """
         Get the size hint.
 
@@ -73,7 +77,7 @@ class ScrollArea(QScrollArea):
 
         Returns
         -------
-        QtCore.QSize
+        QSize
             The size hint for the ScrollArea.
         """
         if self.widget() is not None:
@@ -82,7 +86,7 @@ class ScrollArea(QScrollArea):
         return super().sizeHint()
 
     @Slot()
-    def force_width_from_size_hint(self):
+    def force_width_from_size_hint(self) -> None:
         """
         Enforce a fixed width based on the own sizeHint.
         """