+++ CHANGELOG.rst
@@ -8,6 +8,9 @@ v26.mm.dd
 
 Improvements
 ------------
+- Added an option to set a custom root for data browsing in the 
+  DataBrowsingFrame (and in the UserConfigWindow) to prevent freezes
+  from unresponsive locations
 - Added a dedicated plugin to import NeXus NXdata with full metadata
   and in the shape of the Scan. This allows for conveniently restarting
   workflows from stored intermediate data.
@@ -23,6 +26,8 @@ Bugfixes
   the dialog correctly with the choices.
 - Fixed an issue in the ErrorBox when the error string was too wide to show
   (e.g. for long filenames).
+- Fixed an issue in the DataBrowsingFrame which did not load the icons for 
+  the starting directory correctly.
 
 
 v26.05.19
@@ -39,6 +44,7 @@ Improvements
 - When editing advanced Parameters, the GUI now remembers the settings
   for each plugin.
 - Added an option to select a file by its filename in the DataBrowsingFrame.
+  The field also accepts drag & drop entries.
 - Removed the gui and widget packages from the generic 'import pydidas'
   to speed-up loading for background processes.
 - Added an option in the plots to lock the current zoom settings
@@ -66,6 +72,10 @@ Programmatic changes
 - Added an option to specify the display precision for floating point
   Parameters.
 - Refactored PydidasPlot1D and PydidasPlot2D and sorted methods.
+- Added an option in the UI creation to specify the gridPos row position
+  with "::current::" to append a widget in the current row.
+- Changed the signal of QCheckBox tick state to be consistent with 
+  Qt.QCheckState definitions (checked == 2)
 
 Bugfixes
 --------